bridge/requirements.txt
influxdb_client==1.30.0
paho-mqtt==1.6.1

